requirements.txt
tox
twine
python-logstash==0.4.6
Django>=2.2
